requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai==0.1.7
httpx==0.27.0
# langchain-chroma
chromadb==0.4.24
pypdf
openai==1.30.1